Disable newline if text overflows; revert who/which window width
This is a problem because the game doesn't do auto newlines lazily; that
is, it will greedily scroll the text box on overflow, even if there's
nothing left to be printed.

This behaviour causes a problem when displaying "Which?" in the Equip
screens, because the width of that string is exactly the width of the
window. Because the window is only one line tall, the game was scrolling
down and the string would get erased.

Previously this was worked around by increasing the width of the window,
but recently I have found that it looks weird with the Goods window open.
So this is a slightly different workaround. I don't think we ever wanted
auto-newline-on-overflow _anyway_ since we're putting manual linebreaks
everywhere. Hopefully this doesn't cause problems.