Changed how symbols are resolved
Got rid of the m12-symbols.sym file that I was manually curating; moved
it into m2-hack.asm instead, in the form of .definelabel directives. This
allows existing ROM symbols to be resolved gracefully when the assembler
exports everything, so that we can call existing ROM subroutines, etc.
from our compiled code.

Also added label definitions for hardcoded strings. Makes it easy to refer
to such strings in our compiled code as well.

Updated build scripts to reflect this.