Remove bad movement entry
This was causing the crash before Titanic Ant. My dumb text pointer-
finding script found something in the movement data that it thought was
a text pointer but actually wasn't, and overwrote it with a new pointer
which actually corrupted the movement data!

Should probably remove all of these until we figure out movement codes...